# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation Moon Group",
        "description": "API Documentation for Moon Group",
        "version": "1.0.0"
    },
    "paths": {
        "/moon-group-backend/public/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "operationId": "5a34ca25db7f3aef3d241713a3e1c83d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "username"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "11|SphTzJDxcMxPjpTA3GAMnGMepQKaWMpC05NKn10a1d2879de"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "group_menus": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GroupMenu"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Credentials are invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/moon-group-backend/public/api/authenticate": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user",
                "operationId": "9e74e8c2faf3bc8b2778a5eea124a82c",
                "responses": {
                    "200": {
                        "description": "User authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "11|SphTzJDxcMxPjpTA3GAMnGMepQKaWMpC05NKn10a1d2879de"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "group_menus": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GroupMenu"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moon-group-backend/public/api/logout": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "operationId": "8b0b18ddd887c2e2d39f26b6b6370d72",
                "responses": {
                    "200": {
                        "description": "Logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moon-group-backend/public/api/logs": {
            "get": {
                "tags": [
                    "Logs"
                ],
                "summary": "Get logs",
                "operationId": "d8f3efae3578052e50404d945272429c",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2021-10-01 00:00:00"
                                                    },
                                                    "environment": {
                                                        "type": "string",
                                                        "example": "local"
                                                    },
                                                    "error_type": {
                                                        "type": "string",
                                                        "example": "ERROR"
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Error message"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No logs found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No logs found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GroupMenu": {
                "title": "GroupMenu",
                "required": [
                    "id",
                    "name",
                    "icon"
                ],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "icon": {
                        "type": "string",
                        "example": "fas fa-user"
                    },
                    "order": {
                        "type": "number",
                        "example": "1"
                    },
                    "route": {
                        "type": "string",
                        "example": "slider"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "names": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastnames": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}